Korrekturen

+++ htdocs/admission.inc.php
@@ -40,9 +40,8 @@ function check_admission ($send_message=TRUE) {
 	$db5=new DB_Seminar;
 	$messaging=new messaging;
 	
-	//Alle Veranstaltungen auswaehlen, die in Frage kommen
-	$db->query("SELECT Seminar_id, Name, admission_endtime, admission_turnout, admission_type, start_time FROM seminare WHERE admission_endtime > '".time()."' AND admission_selection_take_place = '0' AND admission_type = '1' '");
-	
+	//Daten holen / Abfrage ob ueberhaupt begrenzt
+	$db->query("SELECT Seminar_id, Name, admission_endtime, admission_turnout, admission_type, start_time FROM seminare WHERE admission_endtime > '".time()."' AND admission_selection_take_place = '0' AND admission_type = '1' ");
 	while ($db->next_record()) {
 		//Veranstaltung locken
 		$db2->query("UPDATE seminare SET admission_selection_take_place ='-1' WHERE Seminar_id = '".$db->f("Seminar_id")."' ");
@@ -51,7 +50,7 @@ function check_admission ($send_message=TRUE) {
 		$db2->query("SELECT studiengang_id, quota FROM admission_seminar_studiengang WHERE seminar_id = '".$db->f("Seminar_id")."' ");
 		while ($db2->next_record()) {
 			//Losfunktion
-			$db3->query("SELECT user_id, username, studiengang_id FROM admission_seminar_user LEFT JOIN auth_user_md5 USING (user_id) WHERE seminar_id = '".$db->f("Seminar_id")."' AND studiengang_id = '".$db2->f("studiengang_id")."' ORDER BY RAND() LIMIT ".round($db->f("admission_turnout") * ($db2->f("quota") / 100)));
+			$db3->query("SELECT admission_seminar_user.user_id, username, studiengang_id FROM admission_seminar_user LEFT JOIN auth_user_md5 USING (user_id) WHERE seminar_id = '".$db->f("Seminar_id")."' AND studiengang_id = '".$db2->f("studiengang_id")."' ORDER BY RAND() LIMIT ".round($db->f("admission_turnout") * ($db2->f("quota") / 100)));
 			//User aus admission_Seminar_user in seminar_user verschieben
 			while ($db3->next_record())   {
 				$group = select_group ($db->f("start_time"), $db3->f("user_id"));			
@@ -94,35 +93,36 @@ function update_admission ($seminar_id, $send_message=TRUE) {
 	$db6=new DB_Seminar;
 	$messaging=new messaging;
 	
-	//Daten holen
-	$db->query("SELECT Seminar_id, Name, admission_endtime, admission_turnout, admission_type, start_time FROM seminare WHERE admission_endtime > '".time()."' AND admission_selection_take_place = '0' AND admission_type = '1' '");
-		
-	//Alle zugelassenen Studiengaenge auswaehlen
-	$db2->query("SELECT studiengang_id, quota FROM admission_seminar_studiengang WHERE seminar_id = '$seminar_id' ");
-	while ($db2->next_record()) {
-		//Wiieviel Teilnehmer sind eingetragen und wieviele koennten eingetragen werden?
-		$db3->query("SELECT count (user_id) AS count FROM seminar_user WHERE Seminar_id = '".$db->f("Seminar_id")."' AND admission_studiengang_id = '".$db2->f("studiengang_id")."' ");
-		$db3->next_record();
-		if ($db3->f("count") < round($db->f("admission_turnout") * ($db2->f("quota") / 100))) {
-			//Studis asuwaehlen, die jetzt aufsteigen koennen
-			$db4->query("SELECT user_id FROM admission_seminar_user WHERE seminar_id =  '".$db->f("Seminar_id")."'  AND studiengang_id = ".$db2->f("studiengang_id")."' ORDER BY position LIMIT ".($db3->f("count") - round($db->f("admission_turnout") * ($db2->f("quota") / 100))));
-			while ($db4->next_record()) {
-				$group = select_group ($db->f("start_time"), $db4->f("user_id"));			
-				$db5->query("INSERT INTO seminar_user SET user_id = '".$db4->f("user_id")."', Seminar_id = '".$db->f("Seminar_id")."', status= 'autor', group = '$group', admission_studiengang_id = '".$db2->f("studiengang_id")."', mkdate = '".time()."' ");
-				if ($db6->num_rows())
-					$db6->query("DELETE FROM admission_seminar_user WHERE user_id ='".$db4->f("user_id")."' AND seminar_id = '".$db->f("Seminar_id")."' ");
-					//User benachrichten
-					if (($db6->num_rows()) && ($send_message)) {
-						$message="Sie als Teilnehmer der Veranstaltung <b>".$db->f("Name")."</b> eingetragen worden, da f&uuml;r Sie ein Platz frei geworden ist. Ab sofort finden Sie die Veranstaltungen in der &Uuml;bersicht ihrer Veranstaltungen. Damit sind sie auch als Teilnehmer der Pr&auml;senzveranstaltung zugelassen.";
-						$messaging->insert_sms ($db3->f("username"), $message);
-					}
-			}
-			//Warteposition der restlichen User neu eintragen
-			$db4->query("SELECT user_id FROM admission_seminar_user WHERE seminar_id =  '".$db->f("Seminar_id")."' ORDER BY position ");
-			$position=1;
-			while ($db4->next_record()) {
-				$db4->query("UPDATE admission_seminar_user SET position = '$position'  ");
-				$position++;
+	//Daten holen / Abfrage ob ueberhaupt begrenzt
+	$db->query("SELECT Seminar_id, Name, admission_endtime, admission_turnout, admission_type, start_time FROM seminare WHERE seminar_id = '$seminar_id' AND admission_type != 0");
+	if ($db->next_record()) {
+		//Alle zugelassenen Studiengaenge auswaehlen
+		$db2->query("SELECT studiengang_id, quota FROM admission_seminar_studiengang WHERE seminar_id = '$seminar_id' ");
+		while ($db2->next_record()) {
+			//Wiieviel Teilnehmer sind eingetragen und wieviele koennten eingetragen werden?
+			$db3->query("SELECT count (user_id) AS count FROM seminar_user WHERE Seminar_id = '".$db->f("Seminar_id")."' AND admission_studiengang_id = '".$db2->f("studiengang_id")."' ");
+			$db3->next_record();
+			if ($db3->f("count") < round($db->f("admission_turnout") * ($db2->f("quota") / 100))) {
+				//Studis asuwaehlen, die jetzt aufsteigen koennen
+				$db4->query("SELECT user_id FROM admission_seminar_user WHERE seminar_id =  '".$db->f("Seminar_id")."'  AND studiengang_id = ".$db2->f("studiengang_id")."' ORDER BY position LIMIT ".($db3->f("count") - round($db->f("admission_turnout") * ($db2->f("quota") / 100))));
+				while ($db4->next_record()) {
+					$group = select_group ($db->f("start_time"), $db4->f("user_id"));			
+					$db5->query("INSERT INTO seminar_user SET user_id = '".$db4->f("user_id")."', Seminar_id = '".$db->f("Seminar_id")."', status= 'autor', group = '$group', admission_studiengang_id = '".$db2->f("studiengang_id")."', mkdate = '".time()."' ");
+					if ($db6->num_rows())
+						$db6->query("DELETE FROM admission_seminar_user WHERE user_id ='".$db4->f("user_id")."' AND seminar_id = '".$db->f("Seminar_id")."' ");
+						//User benachrichten
+						if (($db6->num_rows()) && ($send_message)) {
+							$message="Sie als Teilnehmer der Veranstaltung <b>".$db->f("Name")."</b> eingetragen worden, da f&uuml;r Sie ein Platz frei geworden ist. Ab sofort finden Sie die Veranstaltungen in der &Uuml;bersicht ihrer Veranstaltungen. Damit sind sie auch als Teilnehmer der Pr&auml;senzveranstaltung zugelassen.";
+							$messaging->insert_sms ($db3->f("username"), $message);
+						}
+				}
+				//Warteposition der restlichen User neu eintragen
+				$db4->query("SELECT user_id FROM admission_seminar_user WHERE seminar_id =  '".$db->f("Seminar_id")."' ORDER BY position ");
+				$position=1;
+				while ($db4->next_record()) {
+					$db4->query("UPDATE admission_seminar_user SET position = '$position'  ");
+					$position++;
+				}
 			}
 		}
 	}
